Conflicts:
	src/main/webapp/static/js/views/purchase/orderAdd.js

+++ src/main/java/com/okdeer/jxc/controller/supplier/SupplierAreaController.java
@@ -26,6 +26,7 @@ import com.okdeer.jxc.common.utils.UUIDHexGenerator;
 import com.okdeer.jxc.controller.BaseController;
 import com.okdeer.jxc.supplier.entity.SupplierArea;
 import com.okdeer.jxc.supplier.service.SupplierAreaServiceApi;
+import com.okdeer.jxc.supplier.service.SupplierServiceApi;
 import com.okdeer.jxc.supplier.vo.SupplierAreaVo;
 import com.okdeer.jxc.utils.UserUtil;
 
@@ -51,6 +52,9 @@ public class SupplierAreaController extends
 	
 	@Reference(version = "1.0.0", check = false)
 	private BranchesServiceApi branchesService;
+	
+	@Reference(version = "1.0.0", check = false)
+	private SupplierServiceApi supplierService;
 
 	/**
 	 * @Description: 跳转供应商区域页面
@@ -176,8 +180,12 @@ public class SupplierAreaController extends
 	public RespJson deleteSupplierArea(String areaId) {
 		try {
 			LOG.info("根据区域编码删除供应商区域:{}", areaId);
-			
-			supplierAreaService.deleteSupplierArea(areaId,super.getCurrUserId());
+			int count = supplierService.getSupplierByAreaIdSum(areaId);
+			if(count>0) {
+				return RespJson.error("供应商区域已经被引用！");
+			}else {
+				supplierAreaService.deleteSupplierArea(areaId,super.getCurrUserId());
+			}
 			return RespJson.success();
 		} catch (Exception e) {
 			return RespJson.error("删除供应商区域失败！");
